refactor: Participant 제거 및 상속 관계 삭제
