+++ src/main/java/io/ConsoleOutput.java
@@ -2,20 +2,21 @@ package io;
 
 import bet.BetCenter;
 import game.Card;
-import participant.Participants;
+import participant.Dealer;
 import participant.Player;
+import participant.Players;
 
 import java.util.List;
 import java.util.Map;
 
 public class ConsoleOutput {
 
-    public void printInitialGameSettings(Participants participants) {
-        String joinedPlayers = String.join(", ", participants.getPlayerNames());
+    public void printInitialGameSettings(Players players, Dealer dealer) {
+        String joinedPlayers = String.join(", ", players.getPlayerNames());
         System.out.println("\n딜러와 " + joinedPlayers + "에게 2장을 나누었습니다.");
 
-        System.out.println("딜러카드: " + processCardInfo(participants.openOneCardWithDealer()));
-        for (Player player : participants.getPlayers()) {
+        System.out.println("딜러카드: " + processCardInfo(dealer.openOneCard()));
+        for (Player player : players.getPlayers()) {
             printPlayerCards(player);
         }
     }
@@ -28,10 +29,10 @@ public class ConsoleOutput {
         System.out.println("\n딜러는 16이하라 한장의 카드를 더 받았습니다.");
     }
 
-    public void printGameResults(Participants participants) {
-        System.out.println("\n딜러카드: " + processCardsInfo(participants.openCardsWithDealer()) +
-                " - 결과: " + participants.sumScoreWithDealer());
-        for (Player player : participants.getPlayers()) {
+    public void printGameResults(Players players, Dealer dealer) {
+        System.out.println("\n딜러카드: " + processCardsInfo(dealer.openCards()) +
+                " - 결과: " + dealer.sumCardNumbers());
+        for (Player player : players.getPlayers()) {
             System.out.println(player.getNickname() + "카드: " + processCardsInfo(player.openCards()) + " - 결과: "
                     + player.sumCardNumbers());
         }
@@ -47,10 +48,10 @@ public class ConsoleOutput {
         return card.getNumber().getName() + card.getEmblem().getName();
     }
 
-    public void printFinalProfit(BetCenter betCenter, Participants participants) {
+    public void printFinalProfit(BetCenter betCenter, Dealer dealer) {
         System.out.println("\n## 최종 수익");
-        System.out.println("딜러: " + betCenter.calculateDealerProfit(participants.getDealer()));
-        Map<Player, Integer> deriveBetResults = betCenter.deriveBettingResults(participants.getDealer());
+        System.out.println("딜러: " + betCenter.calculateDealerProfit(dealer));
+        Map<Player, Integer> deriveBetResults = betCenter.deriveBettingResults(dealer);
         for (Player player : deriveBetResults.keySet()) {
             System.out.println(player.getNickname() + ": " + deriveBetResults.get(player));
         }
